requirements.txt
git+https://github.com/Rapptz/discord.py
dnspython==1.16.0
async-timeout==3.0.1
pymongo==3.9.0
pytz==2019.3
discord-webhook==0.5.0
Pillow==6.2.1